Update to remove warnings for OS X 10.6
- Removed all warnings aside from ATSUI deprecation. I need
  to move that over to CoreText.
- Changed encoding of strings files to UTF8 as they were in
  UTF16 and the copy strings build step was expecting UTF8.

+++ Source/Controllers/TriviaPlayerController.m
@@ -53,7 +53,7 @@
 	if( [key isEqualToString:@"players"] && change == NSKeyValueChangeRemoval ) {
 		
 		TriviaPlayer *playerToRemove;
-		unsigned currentIndex = [indexes firstIndex];
+		NSUInteger currentIndex = [indexes firstIndex];
 		while( currentIndex != NSNotFound ) {
 			playerToRemove = [players objectAtIndex:currentIndex];
 			//if( infoController && playerToRemove == [infoController player] )
@@ -154,7 +154,7 @@
 }
 - (IBAction)registerInput:(id)sender
 {	
-	unsigned int selectionIndex = [playerArrayController selectionIndex];
+	NSUInteger selectionIndex = [playerArrayController selectionIndex];
 	if( selectionIndex == NSNotFound || _waitingForButton )
 		return;
